% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_discard.R
\name{obj_discard}
\alias{obj_discard}
\title{Remove objects in global env.}
\usage{
obj_discard(obj_names)
}
\arguments{
\item{obj_names}{Character vector of names as the one given in output by obj_exp.}
}
\value{
Removes objects which names have been passed.
}
\description{
Remove the objects in the global environment which names are passed in.
}
\examples{

obj_discard(obj_names = c("x2", "x1"))

obj_exp(pattern = "x") |> obj_discard()
}
