% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_keep.R
\name{obj_keep}
\alias{obj_keep}
\title{Keeping object in global env.}
\usage{
obj_keep(obj_names)
}
\arguments{
\item{obj_names}{Character vector of names of object to keep.}
}
\value{
Removes object which names have not been passed.
}
\description{
Keep the object in the global environment which names are passed in.
}
\examples{

obj_keep(obj_names = c("x2", "x1"))

obj_exp(pattern = "x") |> obj_keep()
}
