% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_exp.R
\name{obj_exp}
\alias{obj_exp}
\title{Expand a pattern found or not in glabal env objects.}
\usage{
obj_exp(pattern, invert = FALSE)
}
\arguments{
\item{pattern}{A regular expression accepted by grep.}

\item{invert}{Logical value indicating if you want to select the objects that not contains the pattern. Default FALSE.}
}
\value{
A character vector.
}
\description{
Expand a pattern found or not in glabal env objects.
}
\examples{

home11 <- "nice"
home2 <- "ugly"
temperature <- 22

obj_exp(pattern = "home", invert = FALSE)

}
